Add support for adjustable CHT cell size, 32 and 40 currently supported.
